chore: split dev and build tasks
