+++ tasks/build.js
@@ -7,7 +7,6 @@ module.exports = function(gulp, config) {
   ========================================*/
 
   var concat            = require('gulp-concat');
-  var connect           = require('gulp-connect');
   var csso              = require('gulp-csso');
   var del               = require('del');
   var less              = require('gulp-less');
@@ -33,28 +32,6 @@ module.exports = function(gulp, config) {
 
   gulp.task('clean', function() {
     return del(['dist/**']);
-  });
-
-  /*==========================================
-  =            Web servers                   =
-  ==========================================*/
-
-  gulp.task('connect', function() {
-    connect.server({
-      root: process.cwd(),
-      host: '0.0.0.0',
-      port: 3000,
-      livereload: true
-    });
-  });
-
-  /*==============================================================
-  =            Setup live reloading on source changes            =
-  ==============================================================*/
-
-  gulp.task('livereload:demo', function() {
-    return gulp.src(config.globs.livereloadDemo)
-      .pipe(connect.reload());
   });
 
   /*==================================
@@ -143,43 +120,12 @@ module.exports = function(gulp, config) {
 
   gulp.task('js', ['js:main', 'js:gestures', 'js:core']);
 
-  /*===================================================================
-  =            Watch for source changes and rebuild/reload            =
-  ===================================================================*/
-
-  gulp.task('watch', function() {
-    gulp.watch(config.globs.livereloadDemo.concat('dist/**/*'), ['livereload:demo']);
-    gulp.watch(config.globs.livereloadTest.concat('dist/**/*'), ['livereload:test']);
-    gulp.watch(config.globs.livereloadTestManual.concat('dist/**/*'), ['livereload:test:manual']);
-
-    gulp.watch(['./src/less/**/*'], ['css']);
-
-    ['core', 'gestures', 'main'].forEach(function(target) {
-      gulp.watch(config.globs[target], ['js:' + target]);
-    });
-
-    gulp.watch(config.globs.fonts, ['fonts']);
-  });
-
   /*======================================
   =            Build Sequence            =
   ======================================*/
 
   gulp.task('build', function(done) {
     seq('clean', ['fonts', 'css',  'js'], done);
-  });
-
-  /*====================================
-  =            Default Task            =
-  ====================================*/
-
-  gulp.task('default', function(done) {
-    var tasks = [];
-
-    tasks.push('connect');
-    tasks.push('watch');
-
-    seq('build', tasks, done);
   });
 
 };
